// src/app/models/his_pro_model.ts
export default interface HisProModel {
  HP_CODIGO: number;
  HP_DATA: string;
  HP_PRO: number;
  HP_ORIGEM: string;
  HP_DOC: string;
  HP_QUANTIDADE: number;
  HP_VALORC: number;
  HP_VALORV: number;
  HP_VALORCM: number;
  HP_VALOROP: number;
  HP_VALORM: number;
  HP_TIPO: string;
  HP_TIPO2: number;
  HP_QUANTIDADEA: number;
}